requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
idna==2.8
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.0
passlib==1.7.2
paystack==1.5.0
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python-barcode==0.10.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
